feat: add function to brand sheep
when one sheep is clicked and the brand button is pressed,
the sheep is branded.
i.e. theyhave a class applied to them that makes the background green.
the sheep's id is saved to store state for the sheep component to see.
If the id is it's then it adds the class.